backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.0
Pillow==10.3.0
numpy==1.26.4
opencv-python==4.9.0.80
facenet-pytorch==2.5.3
timm==1.0.3
# PyTorch CPU — install manually:
# pip install torch==2.2.2+cpu torchvision==0.17.2+cpu torchaudio==2.2.2+cpu --index-url https://download.pytorch.org/whl/cpu
